+++ crates/project/src/git.rs
@@ -13,9 +13,8 @@ use futures::{
     StreamExt as _,
 };
 use git::{
-    commit::CommitDetails,
     repository::{
-        Branch, GitRepository, PushOptions, Remote, RemoteCommandOutput, RepoPath,
+        Branch, CommitDetails, GitRepository, PushOptions, Remote, RemoteCommandOutput, RepoPath,
         ResetMode,
     },
     status::FileStatus,
